+++ include/envoy/network/BUILD
@@ -35,6 +35,7 @@ envoy_cc_library(
     name = "connection_handler_interface",
     hdrs = ["connection_handler.h"],
     deps = [
+        ":connection_balancer_interface",
         ":listen_socket_interface",
         ":listener_interface",
         "//include/envoy/ssl:context_interface",
