+++ include/envoy/network/connection_handler.h
@@ -4,6 +4,7 @@
 #include <memory>
 
 #include "envoy/network/connection.h"
+#include "envoy/network/connection_balancer.h"
 #include "envoy/network/filter.h"
 #include "envoy/network/listen_socket.h"
 #include "envoy/network/listener.h"
@@ -53,13 +54,6 @@ public:
    * @param listener_tag supplies the tag passed to addListener().
    */
   virtual void removeListeners(uint64_t listener_tag) PURE;
-
-  /**
-   * Get the ``UdpListenerCallbacks`` associated with ``listener_tag``. This will be
-   * absl::nullopt for non-UDP listeners and for ``listener_tag`` values that have already been
-   * removed.
-   */
-  virtual UdpListenerCallbacksOptRef getUdpListenerCallbacks(uint64_t listener_tag) PURE;
 
   /**
    * Remove the filter chains and the connections in the listener. All connections owned
@@ -161,6 +155,24 @@ public:
 
 using ConnectionHandlerPtr = std::unique_ptr<ConnectionHandler>;
 
+class TcpConnectionHandler : public virtual ConnectionHandler {
+public:
+  virtual Event::Dispatcher& dispatcher() PURE;
+  virtual BalancedConnectionHandlerOptRef getBalancedHandlerByTag(uint64_t listener_tag) PURE;
+  virtual BalancedConnectionHandlerOptRef
+  getBalancedHandlerByAddress(const Network::Address::Instance& address) PURE;
+};
+
+class UdpConnectionHandler : public virtual ConnectionHandler {
+public:
+  /**
+   * Get the ``UdpListenerCallbacks`` associated with ``listener_tag``. This will be
+   * absl::nullopt for non-UDP listeners and for ``listener_tag`` values that have already been
+   * removed.
+   */
+  virtual UdpListenerCallbacksOptRef getUdpListenerCallbacks(uint64_t listener_tag) PURE;
+};
+
 /**
  * A registered factory interface to create different kinds of ActiveUdpListener.
  */
@@ -179,7 +191,7 @@ public:
    * @return the ActiveUdpListener created.
    */
   virtual ConnectionHandler::ActiveUdpListenerPtr
-  createActiveUdpListener(uint32_t worker_index, ConnectionHandler& parent,
+  createActiveUdpListener(uint32_t worker_index, UdpConnectionHandler& parent,
                           Event::Dispatcher& dispatcher, Network::ListenerConfig& config) PURE;
 
   /**
